removing gandalf_completion
