+++ pyrit/completion/__init__.py
@@ -2,6 +2,5 @@
 # Licensed under the MIT license.
 
 from pyrit.completion.azure_completions import AzureCompletion
-from pyrit.completion.gandalf_completion import GandalfCompletionEngine, GandalfLevel
 
-__all__ = ["AzureCompletion", "GandalfCompletionEngine", "GandalfLevel"]
+__all__ = ["AzureCompletion"]
