+++ pyrit/prompt_target/gandalf_target.py
@@ -3,10 +3,9 @@
 
 import asyncio
 import concurrent.futures
+import enum
 import logging
-from multiprocessing import pool
 
-from pyrit.completion import GandalfLevel
 from pyrit.common import net_utility
 from pyrit.memory import DuckDBMemory, MemoryInterface
 from pyrit.models import PromptRequestResponse
@@ -17,6 +16,19 @@ logger = logging.getLogger(__name__)
 
 
 class GandalfTarget(PromptTarget):
+
+    class GandalfLevel(enum.Enum):
+        LEVEL_1 = "baseline"
+        LEVEL_2 = "do-not-tell"
+        LEVEL_3 = "do-not-tell-and-block"
+        LEVEL_4 = "gpt-is-password-encoded"
+        LEVEL_5 = "word-blacklist"
+        LEVEL_6 = "gpt-blacklist"
+        LEVEL_7 = "gandalf"
+        LEVEL_8 = "gandalf-the-white"
+        LEVEL_9 = "adventure-1"
+        LEVEL_10 = "adventure-2"
+
     def __init__(
         self,
         *,
@@ -35,7 +47,6 @@ class GandalfTarget(PromptTarget):
         pool = concurrent.futures.ThreadPoolExecutor()
         return pool.submit(asyncio.run, self.send_prompt_async(prompt_request=prompt_request)).result()
 
-
     async def send_prompt_async(self, *, prompt_request: PromptRequestResponse) -> PromptRequestResponse:
         request = prompt_request.request_pieces[0]
 
@@ -48,9 +59,7 @@ class GandalfTarget(PromptTarget):
 
         response = await self._complete_text_async(request.converted_prompt_text)
 
-        response_entry = self._memory.add_response_entries_to_memory(
-            request=request, response_text_pieces=[response]
-        )
+        response_entry = self._memory.add_response_entries_to_memory(request=request, response_text_pieces=[response])
 
         return response_entry
